+++ Include/cpython/pystate.h
@@ -6,6 +6,8 @@
 extern "C" {
 #endif
 
+#include "coreconfig.h"
+
 /* Placeholders while working on the new configuration API
  *
  * See PEP 432 for final anticipated contents
